Fixed edit image

+++ userRoutes.ts
@@ -415,12 +415,18 @@ userRoutes.put("/editUserInfo", upload.single("image"), async function (
     //after checking user name, del image if new image is uploaded
     const currentImage: string = currentUser.image;
 
+    console.log("currentImage: " + currentImage);
+
+    console.log("req.file: " + req.file);
+
     let image: string | "";
     if (req.file) {
       image = req.file.filename;
-      await fs.unlinkSync("./public/uploads/" + currentImage);
+      if (currentImage) {
+        await fs.unlinkSync("./public/uploads/" + currentImage);
+      }
     } else {
-      image = "";
+      image = currentImage;
     }
 
     //insert user into sql
